+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Return the first and deepest issue detected. ([#317](https://github.com/getsentry/vroom/pull/317))
 
+**Bug Fixes**
+
+- Close remaining open events in Android profiles. ([#316](https://github.com/getsentry/vroom/pull/316))
+
 **Internal**
 
 - Rename the frame drop issue title. ([#315](https://github.com/getsentry/vroom/pull/315))